Packaging part of sofa project.
An Inspiration and Modification from
github.com/pjw91/sofa/tree/TEST-will-always-be-force-pushed

Since it's more handy and proper for user to install sofa which is packaged.
Plus, as a user, I install sofa manually that ```tools/prepare.sh``` and ```install.sh```
isn't suitable for my environment and OS distro.

The steps of installation will become:
1. Build Dependencies (e.g., sudo apt install gcc) manually
2. Install sofa with ```python3 -m pip install .```

NOTICE that sofa-pcm isn't build for this commit.

Remove origin bin/* and modified .travis.yml